#1: Reorganizing code in multiple files
