--- src/server/python/__init__.py
@@ -1,27 +1,0 @@
-# __init__.py
-#
-# Copyright (C) 2024 Felix GAIDON
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-"""
-This package contains the FastAPI application and related test scripts.
-
-The package serves as part of the 'daily-quote' project, providing the
-backend logic for posting and reading quotes. The application is developed
-to explore the implementation of a distributed system.
-"""
-
-# By default, this file is left minimal as it marks this directory as a package.
-# Additional code or initialization logic can be added here as needed for
-# future use.
